squelch lots of warnings (MSVC) and corrected some oversized types
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@9774 8aca7d54-2c30-db11-9de9-000461428c89

# Conflicts:
#	Simutrans-GDI.vcxproj
#	Simutrans-SDL2.vcxproj
#	Simutrans-Server.vcxproj
#	bauer/fabrikbauer.cc
#	boden/wasser.cc
#	dataobj/schedule.cc
#	dataobj/settings.h
#	gui/climates.cc
#	gui/components/gui_schedule.cc
#	gui/display_settings.cc
#	simconvoi.cc
#	simhalt.cc
#	simmain.cc
#	simmenu.cc
#	simtool-dialogs.h
#	simtool.cc
#	sound/win32_sound_xa.cc
#	utils/simrandom.cc
#	vehicle/air_vehicle.cc

+++ simconvoi.cc
@@ -5181,9 +5181,8 @@ void convoi_t::get_freight_info(cbuffer_t & buf)
 
 		// apend info on total capacity
 		slist_tpl <ware_t>capacity;
-
-		for (size_t i = 0; i != n; ++i) {
-			if (max_loaded_waren[i] > 0 && i != goods_manager_t::INDEX_NONE) {
+		for (uint16 i = 0; i != n; ++i) {
+			if(max_loaded_waren[i]>0  &&  i!=goods_manager_t::INDEX_NONE) {
 				ware_t ware(goods_manager_t::get_info(i));
 				ware.menge = max_loaded_waren[i];
 				// append to category?
